backend/requirements.txt
flask
google-generativeai
pandas
gunicorn